+++ docs/source/conf.py
@@ -49,7 +49,6 @@ exclude_patterns = [
     "utils",
     "inferers",
     "optimizers",
-    "auto3dseg",
 ]
