+++ monai/__init__.py
@@ -49,7 +49,6 @@ load_submodules(sys.modules[__name__], True, exclude_pattern=excludes)
 
 __all__ = [
     "apps",
-    "auto3dseg",
     "bundle",
     "config",
     "data",
